+++ src/constants/tunnel.ts
@@ -1,0 +1,12 @@
+/**
+ * @file Tunnel constant
+ * @module constant/tunnel
+ * @author Surmon <https://github.com/surmon-china>
+ */
+
+export enum TunnelModule {
+  Bilibili = 'bilibili',
+  Wallpaper = 'wallpaper',
+  GitHub = 'gitHub',
+  Music = 'music'
+}